Refactor filter requires to use relative paths and remove obsolete logging methods in trace data classes
