+++ lib/trace_viz/formatters/loggers/method_call_formatter.rb
@@ -1,19 +1,18 @@
 # frozen_string_literal: true
 
-require "trace_viz/formatters/trace_point/base_formatter"
+require_relative "base_formatter"
 
 module TraceViz
   module Formatters
-    module TracePoint
-      class MethodReturnFormatter < Formatters::BaseFormatter
-        def format_for_log
+    module Loggers
+      class MethodCallFormatter < BaseFormatter
+        def format
           [
             indent_if_enabled,
             colorize(depth_if_enabled, :blue),
             colorize(method_name_if_enabled, :light_green),
-            colorize(result_if_enabled, :cyan),
             colorize(source_location_if_enabled, :dark_gray),
-            colorize(execution_time_if_enabled, :red),
+            colorize(params_if_enabled, :yellow),
           ].compact.join(" ")
         end
       end
