+++ lib/trace_viz/formatters/loggers/method_return_formatter.rb
@@ -1,18 +1,19 @@
 # frozen_string_literal: true
 
-require "trace_viz/formatters/trace_point/base_formatter"
+require_relative "base_formatter"
 
 module TraceViz
   module Formatters
-    module TracePoint
-      class MethodCallFormatter < Formatters::BaseFormatter
-        def format_for_log
+    module Loggers
+      class MethodReturnFormatter < BaseFormatter
+        def format
           [
             indent_if_enabled,
             colorize(depth_if_enabled, :blue),
             colorize(method_name_if_enabled, :light_green),
+            colorize(result_if_enabled, :cyan),
             colorize(source_location_if_enabled, :dark_gray),
-            colorize(params_if_enabled, :yellow),
+            colorize(execution_time_if_enabled, :red),
           ].compact.join(" ")
         end
       end
