+++ lib/trace_viz/trace_data/trace_point/method_call.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require "trace_viz/trace_data/trace_point/base"
-require "trace_viz/formatters/trace_point/method_call_formatter"
+require_relative "base"
+require "trace_viz/formatters/loggers/method_call_formatter"
 
 module TraceViz
   module TraceData
@@ -10,23 +10,6 @@ module TraceViz
         def initialize(trace_point)
           super
           increment_depth
-        end
-
-        def log_trace
-          return unless should_log?
-
-          formatted_call = Formatters::TracePoint::MethodCallFormatter.new(self).format_for_log
-
-          logger.start(formatted_call)
-        end
-
-        private
-
-        def increment_depth
-          return @depth = 0 unless tracker&.depth
-
-          @depth = tracker.depth.current || 0
-          tracker.depth.increment
         end
       end
     end
