+++ lib/trace_viz/trace_data/trace_point/method_return.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require "trace_viz/trace_data/trace_point/base"
-require "trace_viz/formatters/trace_point/method_return_formatter"
+require_relative "base"
+require "trace_viz/formatters/loggers/method_return_formatter"
 
 module TraceViz
   module TraceData
@@ -10,22 +10,6 @@ module TraceViz
         def initialize(trace_point)
           super
           decrement_depth
-        end
-
-        def log_trace
-          return unless should_log?
-
-          formatted_return = Formatters::TracePoint::MethodReturnFormatter.new(self).format_for_log
-
-          logger.finish(formatted_return)
-        end
-
-        private
-
-        def decrement_depth
-          return 0 unless tracker&.depth
-
-          @depth = tracker.depth.decrement
         end
       end
     end
